Added Classification methods
Added functions for performing classification with DIRAC, RACE, and
CRANE. Added main method for comparing phenotypes based on rank entropy
to simplify interface. Updated Rogygen tags with includes, so that
needed dependencies are loaded correctly, and regenerated documentation
(this is the reason for the large number of updated files). Added empty
vignette for future use.